Create extension point to package pregenerated skeletons
