+++ python/src/com/jetbrains/python/sdk/skeletons/PyPregeneratedSkeletons.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2000-2017 JetBrains s.r.o.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.jetbrains.python.sdk.skeletons;
+
+import com.jetbrains.python.sdk.InvalidSdkException;
+
+
+public interface PyPregeneratedSkeletons {
+  boolean copyPregeneratedSkeleton(String moduleName, String skeletonDir) throws InvalidSdkException;
+
+  void unpackPreGeneratedSkeletons(String skeletonDir) throws InvalidSdkException;
+}
